cleanup and completion through map class end
restructured files to be consistent, .h files all contain constructors/destructors and have public members first and private toward end of file. map's export function now appends the data to a map.exportBuffer adn export only purges and writes to a temp file when the buffer size is met or if we reached EOF for that particular file and must purge the buffer or remaining tokens.